requirements.txt
django==1.4.5
django-category==0.1
django-debug-toolbar
django-export
django-generate
django-google-analytics-app
django-google-credentials
django-haystack==1.2.7
django-preferences
django-sites-groups
django-snippetscream
django-south-admin
django-registration==0.8
django-userprofile>=0.0.6
django-registration==0.8
django-ckeditor
gunicorn
jmbo==0.5.5
jmbo-poll>=0.0.9
jmbo-post
celery==3.1.10
django-celery==3.1.10
kombu>3.0
photon
psycopg2
python-ambient
python-dateutil
python-memcached
raven
south
Whoosh==2.4.1
django-celery-email

django-tastypie < 0.10.0
mimeparse
django-moderator
requests
xmltodict
https://github.com/praekelt/django-photologue/tarball/2.10.praekelt#egg=django-photologue
raven

-e git+git://github.com/praekelt/jmbo-your-words.git@develop#egg=jmbo-your-words
-e git+git://github.com/praekelt/mama-survey.git@develop#egg=mama-survey
-e git+git://github.com/praekelt/django-livechat.git@develop#egg=django-livechat

mock
nose
django-nose
coverage
